#58: post-review refactoring
