+++ server/src/main/scala/za/co/absa/atum/server/api/repository/CheckpointRepository.scala
@@ -17,7 +17,6 @@
 package za.co.absa.atum.server.api.repository
 
 import za.co.absa.atum.model.dto.CheckpointDTO
-import za.co.absa.atum.server.api.database.runs.functions.WriteCheckpoint
 import za.co.absa.atum.server.api.exception.DatabaseError
 import za.co.absa.fadb.exceptions.StatusException
 import zio._
@@ -26,20 +25,4 @@ import zio.macros.accessible
 @accessible
 trait CheckpointRepository {
   def writeCheckpoint(checkpointDTO: CheckpointDTO): IO[DatabaseError, Either[StatusException, Unit]]
-}
-
-class CheckpointRepositoryImpl(writeCheckpointFn: WriteCheckpoint) extends CheckpointRepository with BaseRepository {
-
-  override def writeCheckpoint(checkpointDTO: CheckpointDTO): IO[DatabaseError, Either[StatusException, Unit]] = {
-    dbCallWithStatus(writeCheckpointFn.apply, checkpointDTO)
-  }
-
-}
-
-object CheckpointRepositoryImpl {
-  val layer: URLayer[WriteCheckpoint, CheckpointRepository] = ZLayer {
-    for {
-      writeCheckpoint <- ZIO.service[WriteCheckpoint]
-    } yield new CheckpointRepositoryImpl(writeCheckpoint)
-  }
 }
